fun times with wall times

+++ pkg/cache/scheduler/cache.go
@@ -707,7 +707,7 @@ type ClusterQueueUsageStats struct {
 	ReservingWorkloads  int
 	AdmittedResources   []kueue.FlavorUsage
 	AdmittedWorkloads   int
-	WeightedShare       int64
+	WeightedShare       float64
 	WallTimeFlavorUsage []kueue.WallTimeFlavorUsage
 }
